backend/requirements.txt
# FastAPI & Server
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pgvector==0.3.4
alembic==1.13.2

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# AI / LLM
langchain==0.3.7
langchain-openai==0.2.6
langchain-community==0.3.5
openai==1.54.0

# File Processing
pdfplumber==0.11.4
python-docx==1.1.2

# Utilities
python-dotenv==1.0.1
pydantic[email]==2.9.2
pydantic-settings==2.5.2
numpy==1.26.4
httpx==0.27.2
